+++ Cargo.toml
@@ -84,4 +84,4 @@ dao-interface = { path = "./packages/dao-interface" }
 dao-voting = { path = "./packages/dao-voting" }
 cw-proposal-single-v1 = { package = "cw-proposal-single", version = "0.1.0", git = "https://github.com/DA0-DA0/dao-contracts.git", tag = "v1.0.0" }
 voting-v1 = { package = "voting", version = "0.1.0", git = "https://github.com/DA0-DA0/dao-contracts.git", tag = "v1.0.0" }
-cw-utils-v1 = {package = "cw-utils", version = "0.13"}
+cw-utils-v1 = { package = "cw-utils", version = "0.13" }
