+++ contracts/external/cw-payroll/src/contract.rs
@@ -1,24 +1,21 @@
-use std::vec;
-
-use crate::balance::*;
-use crate::error::{ContractError, GenericError};
-use crate::linking::*;
-use crate::msg::{
-    ConfigResponse, ContractResult, ExecuteMsg, InstantiateMsg, ListStreamsResponse, QueryMsg,
-    ReceiveMsg, StreamId, StreamParams, StreamResponse,
-};
-use crate::state::{
-    add_stream, remove_stream, save_stream, Config, Stream, CONFIG, STREAMS, STREAM_SEQ,
-};
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{
-    from_binary, to_binary, BankMsg, Binary, Coin, CosmosMsg, Deps, DepsMut, Env, MessageInfo,
-    Order, Response, StdResult, Storage, Uint128,
+    from_binary, to_binary, Binary, Deps, DepsMut, Env, MessageInfo, Order, Response, StdError,
+    StdResult, Storage, Uint128,
 };
 use cw2::set_contract_version;
-use cw20::{Cw20CoinVerified, Cw20Contract, Cw20ExecuteMsg, Cw20ReceiveMsg};
+use cw20::Cw20ReceiveMsg;
+use cw_denom::UncheckedDenom;
 use cw_storage_plus::Bound;
+
+use crate::error::ContractError;
+use crate::linking::*;
+use crate::msg::{
+    ConfigResponse, ExecuteMsg, InstantiateMsg, ListStreamsResponse, QueryMsg, ReceiveMsg,
+    StreamParams, StreamResponse,
+};
+use crate::state::{Config, Stream, StreamId, CONFIG, STREAMS, STREAM_SEQ};
 
 const CONTRACT_NAME: &str = "crates.io:cw-escrow-streams";
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -29,7 +26,7 @@ pub fn instantiate(
     _env: Env,
     info: MessageInfo,
     msg: InstantiateMsg,
-) -> ContractResult {
+) -> Result<Response, ContractError> {
     set_contract_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
 
     let admin = match msg.admin {
@@ -44,15 +41,24 @@ pub fn instantiate(
 
     STREAM_SEQ.save(deps.storage, &0u64)?;
 
+    // TODO optionally fund on instantiate?
+
     Ok(Response::new()
         .add_attribute("method", "instantiate")
         .add_attribute("admin", admin))
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
-pub fn execute(deps: DepsMut, env: Env, info: MessageInfo, msg: ExecuteMsg) -> ContractResult {
+pub fn execute(
+    deps: DepsMut,
+    env: Env,
+    info: MessageInfo,
+    msg: ExecuteMsg,
+) -> Result<Response, ContractError> {
     match msg {
         ExecuteMsg::Receive(msg) => execute_receive(env, deps, info, msg),
+        // TODO should be able to create and fund with a native token
+        ExecuteMsg::Create {} => unimplemented!(),
         ExecuteMsg::Distribute { id } => execute_distribute(env, deps, id),
         ExecuteMsg::PauseStream { id } => execute_pause_stream(env, deps, info, id),
         ExecuteMsg::ResumeStream { id } => execute_resume_stream(env, deps, info, id),
@@ -67,7 +73,7 @@ pub fn execute_pause_stream(
     deps: DepsMut,
     info: MessageInfo,
     id: StreamId,
-) -> ContractResult {
+) -> Result<Response, ContractError> {
     let pause_stream_local =
         |stream_id: StreamId, storage: &mut dyn Storage| -> Result<Stream, ContractError> {
             let mut stream = STREAMS
@@ -81,7 +87,7 @@ pub fn execute_pause_stream(
             }
             stream.paused_time = Some(env.block.time.seconds());
             stream.paused = true;
-            save_stream(storage, id, &stream).unwrap();
+            STREAMS.save(storage, id, &stream)?;
             Ok(stream)
         };
     let stream = pause_stream_local(id, deps.storage).unwrap();
@@ -98,64 +104,46 @@ pub fn execute_pause_stream(
 
     Ok(response)
 }
+
 pub fn execute_remove_stream(
     env: Env,
     deps: DepsMut,
     info: MessageInfo,
     id: StreamId,
-) -> ContractResult {
+) -> Result<Response, ContractError> {
+    // Check that sender is admin
     let stream = STREAMS
         .may_load(deps.storage, id)?
         .ok_or(ContractError::StreamNotFound { stream_id: id })?;
-
     if stream.admin != info.sender {
         return Err(ContractError::Unauthorized {});
     }
+
+    // TODO what do we want here?
     if let Some(link_id) = stream.link_id {
         return Err(ContractError::LinkedStreamDeleteNotAllowed { link_id });
     }
-    remove_stream(deps.storage, id).unwrap();
-    let response = Response::new()
+    STREAMS.remove(deps.storage, id);
+
+    // Transfer any remaining balance to the owner
+    let transfer_to_admin_msg = stream
+        .denom
+        .get_transfer_to_message(&stream.admin, stream.balance)?;
+
+    Ok(Response::new()
         .add_attribute("method", "remove_stream")
         .add_attribute("stream_id", id.to_string())
         .add_attribute("admin", info.sender)
         .add_attribute("removed_time", env.block.time.to_string())
-        .add_message(
-            create_balance_transfer_msg(&stream.balance, None, stream.admin.clone().into())
-                .unwrap(),
-        );
-    Ok(response)
+        .add_message(transfer_to_admin_msg))
 }
-fn create_balance_transfer_msg(
-    balance: &WrappedBalance,
-    amount: Option<Uint128>,
-    recepient: String,
-) -> Result<CosmosMsg, ContractError> {
-    if let Some(native_balance) = balance.native() {
-        let msg = CosmosMsg::Bank(BankMsg::Send {
-            to_address: recepient,
-            amount: vec![Coin {
-                denom: native_balance.denom.clone(),
-                amount: amount.unwrap_or(native_balance.amount),
-            }],
-        });
-        return Ok(msg);
-    } else if let Some(cw20_balance) = balance.cw20() {
-        let cw20 = Cw20Contract(cw20_balance.address.clone());
-        let msg = cw20.call(Cw20ExecuteMsg::Transfer {
-            recipient: recepient,
-            amount: amount.unwrap_or(cw20_balance.amount),
-        })?;
-        return Ok(msg);
-    }
-    Err(ContractError::CouldNotCreateBankMessage {})
-}
+
 pub fn execute_resume_stream(
     env: Env,
     deps: DepsMut,
     info: MessageInfo,
     id: StreamId,
-) -> ContractResult {
+) -> Result<Response, ContractError> {
     let resume_stream_local =
         |stream_id: StreamId, storage: &mut dyn Storage| -> Result<Stream, ContractError> {
             let mut stream = STREAMS
@@ -170,7 +158,7 @@ pub fn execute_resume_stream(
             stream.paused_duration = stream.calc_pause_duration(env.block.time);
             stream.paused = false;
             stream.paused_time = None;
-            save_stream(storage, id, &stream).unwrap();
+            STREAMS.save(storage, id, &stream)?;
             Ok(stream)
         };
 
@@ -193,11 +181,17 @@ pub fn execute_resume_stream(
 
     Ok(response)
 }
-pub fn execute_create_stream(env: Env, deps: DepsMut, params: StreamParams) -> ContractResult {
+
+pub fn execute_create_stream(
+    env: Env,
+    deps: DepsMut,
+    params: StreamParams,
+) -> Result<Response, ContractError> {
     let StreamParams {
         admin,
         recipient,
         balance,
+        denom,
         start_time,
         end_time,
         title,
@@ -218,27 +212,12 @@ pub fn execute_create_stream(env: Env, deps: DepsMut, params: StreamParams) -> C
         return Err(ContractError::InvalidEndTime {});
     }
 
-    let duration: u128 = (end_time - start_time).into();
-    let balance_amount = balance.amount();
-    let refund: u128 = balance_amount
-        .checked_rem(duration)
-        .ok_or(GenericError::IntegerOverflow {})
-        .unwrap();
-
-    if balance_amount < duration {
-        return Err(ContractError::AmountLessThanDuration {});
-    }
-
-    let claimed = if balance.is_native() {
-        WrappedBalance::default()
-    } else {
-        WrappedBalance::new_cw20(balance.cw20().unwrap().address.clone(), Uint128::new(0))
-    };
     let stream = Stream {
         admin: admin.clone(),
         recipient: recipient.clone(),
-        balance: balance.clone(),
-        claimed_balance: claimed,
+        balance,
+        claimed_balance: Uint128::zero(),
+        denom,
         start_time,
         end_time,
         paused_time: None,
@@ -249,34 +228,32 @@ pub fn execute_create_stream(env: Env, deps: DepsMut, params: StreamParams) -> C
         link_id: None,
         is_detachable: is_detachable.unwrap_or(true),
     };
-    let id = add_stream(deps, &stream)?;
 
-    let mut response = Response::new()
+    let id = STREAM_SEQ.load(deps.storage)?;
+    let id = id + 1;
+    STREAM_SEQ.save(deps.storage, &id)?;
+    STREAMS.save(deps.storage, id, &stream)?;
+
+    Ok(Response::new()
         .add_attribute("method", "create_stream")
         .add_attribute("stream_id", id.to_string())
         .add_attribute("admin", admin.clone())
         .add_attribute("recipient", recipient)
         .add_attribute("start_time", start_time.to_string())
-        .add_attribute("end_time", end_time.to_string());
-
-    if refund > 0 {
-        response = response.add_messages(create_balance_transfer_msg(
-            &balance,
-            Some(Uint128::new(refund)),
-            admin.into(),
-        ));
-    }
-    Ok(response)
+        .add_attribute("end_time", end_time.to_string()))
 }
 
 pub fn execute_receive(
     env: Env,
     deps: DepsMut,
     info: MessageInfo,
-    wrapped: Cw20ReceiveMsg,
-) -> ContractResult {
-    deps.api.addr_validate(&info.sender.into_string())?;
-    let msg: ReceiveMsg = from_binary(&wrapped.msg)?;
+    receive_msg: Cw20ReceiveMsg,
+) -> Result<Response, ContractError> {
+    deps.api.addr_validate(&info.sender.clone().into_string())?;
+    let msg: ReceiveMsg = from_binary(&receive_msg.msg)?;
+    let checked_denom =
+        UncheckedDenom::Cw20(info.sender.to_string()).into_checked(deps.as_ref())?;
+
     match msg {
         ReceiveMsg::CreateStream {
             admin,
@@ -288,9 +265,10 @@ pub fn execute_receive(
             env,
             deps,
             StreamParams {
-                admin: admin.unwrap_or_else(|| wrapped.sender.clone()),
+                admin: admin.unwrap_or_else(|| receive_msg.sender.clone()),
                 recipient,
-                balance: WrappedBalance::from(wrapped),
+                balance: receive_msg.amount,
+                denom: checked_denom,
                 start_time,
                 end_time,
                 title: None,
@@ -301,12 +279,11 @@ pub fn execute_receive(
     }
 }
 
-pub fn execute_distribute(env: Env, deps: DepsMut, id: u64) -> ContractResult {
-    let mut stream = STREAMS
+pub fn execute_distribute(env: Env, deps: DepsMut, id: u64) -> Result<Response, ContractError> {
+    let stream = STREAMS
         .may_load(deps.storage, id)?
         .ok_or(ContractError::StreamNotFound { stream_id: id })?;
 
-    let mut msgs: Vec<CosmosMsg> = vec![];
     let (available_claims, _) = stream.calc_distribution_rate(env.block.time);
 
     if !stream.can_distribute_more() || available_claims.u128() == 0 {
@@ -314,54 +291,31 @@ pub fn execute_distribute(env: Env, deps: DepsMut, id: u64) -> ContractResult {
             claimed: stream.claimed_balance,
         });
     }
-    if let Some(native_balance) = stream.balance.native() {
-        let to_claim = Coin {
-            denom: native_balance.denom.clone(),
-            amount: available_claims,
-        };
 
-        stream
-            .claimed_balance
-            .checked_add_native(&[to_claim.clone()])
-            .unwrap();
-        let msg = CosmosMsg::Bank(BankMsg::Send {
-            to_address: stream.recipient.clone().into(),
-            amount: vec![Coin {
-                denom: native_balance.denom.clone(),
-                amount: to_claim.amount,
-            }],
-        });
-        msgs.push(msg);
-        stream.balance.checked_sub_native(&[to_claim]).unwrap();
-    } else if let Some(cw20) = stream.balance.cw20() {
-        let to_claim = Cw20CoinVerified {
-            address: cw20.address.clone(),
-            amount: available_claims,
-        };
-        stream
-            .claimed_balance
-            .checked_add_cw20(&[to_claim.clone()])
-            .unwrap();
+    // Update claimed amount
+    stream
+        .claimed_balance
+        .checked_add(available_claims)
+        .map_err(StdError::overflow)?;
+    // Update remaining balance
+    stream
+        .balance
+        .checked_sub(available_claims)
+        .map_err(StdError::overflow)?;
 
-        let cw20 = Cw20Contract(cw20.address.clone());
-        let msg = cw20.call(Cw20ExecuteMsg::Transfer {
-            recipient: stream.recipient.clone().into(),
-            amount: to_claim.amount,
-        })?;
-        msgs.push(msg);
-        stream.balance.checked_sub_cw20(&[to_claim]).unwrap();
-    }
+    // Save updated stream
+    STREAMS.save(deps.storage, id, &stream)?;
 
-    save_stream(deps.storage, id, &stream).unwrap();
+    // Get transfer message
+    let transfer_msg = stream
+        .denom
+        .get_transfer_to_message(&stream.recipient, available_claims)?;
 
-    let mut res = Response::new()
+    Ok(Response::new()
         .add_attribute("method", "distribute")
         .add_attribute("vested", available_claims)
-        .add_attribute("stream_id", id.to_string());
-    if !msgs.is_empty() {
-        res = res.add_messages(msgs);
-    }
-    Ok(res)
+        .add_attribute("stream_id", id.to_string())
+        .add_message(transfer_msg))
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
@@ -390,6 +344,7 @@ fn query_stream(deps: Deps, id: u64) -> StdResult<StreamResponse> {
         recipient: stream.recipient.into(),
         balance: stream.balance,
         claimed_balance: stream.claimed_balance,
+        denom: stream.denom,
         start_time: stream.start_time,
         end_time: stream.end_time,
         title: stream.title,
@@ -425,6 +380,7 @@ fn map_stream(item: StdResult<(u64, Stream)>) -> StdResult<StreamResponse> {
         recipient: stream.recipient.to_string(),
         balance: stream.balance,
         claimed_balance: stream.claimed_balance,
+        denom: stream.denom,
         start_time: stream.start_time,
         end_time: stream.end_time,
         title: stream.title,
@@ -435,1028 +391,4 @@ fn map_stream(item: StdResult<(u64, Stream)>) -> StdResult<StreamResponse> {
         is_detachable: stream.is_detachable,
         link_id: stream.link_id,
     })
-}
-
-#[cfg(test)]
-mod tests {
-
-    use crate::msg::StreamIdsExtensions;
-
-    use super::*;
-    use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
-    use cosmwasm_std::{Addr, CosmosMsg, StdError, WasmMsg};
-
-    fn get_stream(deps: Deps, id: u64) -> StdResult<Stream> {
-        let msg = QueryMsg::GetStream { id };
-        let res = query(deps, mock_env(), msg)?;
-        from_binary::<Stream>(&res)
-    }
-
-    #[test]
-    fn test_initialization() {
-        let mut deps = mock_dependencies();
-        let msg = InstantiateMsg { admin: None };
-
-        let info = mock_info("creator", &[]);
-        instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let msg = QueryMsg::GetConfig {};
-        let res = query(deps.as_ref(), mock_env(), msg).unwrap();
-        let config: Config = from_binary(&res).unwrap();
-
-        assert_eq!(
-            config,
-            Config {
-                admin: Addr::unchecked("creator")
-            }
-        );
-    }
-
-    #[test]
-    fn test_execute_distribute() {
-        let mut deps = mock_dependencies();
-        let msg = InstantiateMsg { admin: None };
-        let info = mock_info("cw20", &[]);
-        instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let admin_addr = Addr::unchecked("cw20");
-        let sender = admin_addr.to_string();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(1000);
-
-        let balance = WrappedBalance::new_cw20(Addr::unchecked("cw20"), amount);
-        let claimed = WrappedBalance::new_cw20(Addr::unchecked("cw20"), Uint128::new(0));
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(300).seconds();
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), env.clone(), info, msg).unwrap();
-
-        assert_eq!(
-            get_stream(deps.as_ref(), 1).unwrap(),
-            Stream {
-                admin: admin_addr.clone(),
-                recipient: Addr::unchecked("bob"),
-                balance,
-                claimed_balance: claimed.clone(),
-                start_time,
-                end_time,
-                title: None,
-                description: None,
-                paused: false,
-                paused_time: None,
-                paused_duration: None,
-                link_id: None,
-                is_detachable: true,
-            }
-        );
-
-        // Stream has not started
-        let mut info = mock_info("owner", &[]);
-        info.sender = Addr::unchecked("bob");
-        let msg = ExecuteMsg::Distribute { id: 1 };
-        let err = execute(deps.as_mut(), env, info, msg).unwrap_err();
-        assert_eq!(err, ContractError::NoFundsToClaim { claimed });
-
-        // Stream has started so tokens have vested
-        let msg = ExecuteMsg::Distribute { id: 1 };
-        let mut info = mock_info("owner", &[]);
-        let mut env = mock_env();
-        let sender = Addr::unchecked("bob");
-        info.sender = sender;
-        env.block.time = env.block.time.plus_seconds(150);
-        let res = execute(deps.as_mut(), env, info, msg).unwrap();
-        let msg = res.messages[0].clone().msg;
-
-        assert_eq!(
-            msg,
-            CosmosMsg::Wasm(WasmMsg::Execute {
-                contract_addr: String::from("cw20"),
-                msg: to_binary(&Cw20ExecuteMsg::Transfer {
-                    recipient: String::from("bob"),
-                    amount: Uint128::new(250)
-                })
-                .unwrap(),
-                funds: vec![]
-            })
-        );
-        assert_eq!(
-            get_stream(deps.as_ref(), 1).unwrap(),
-            Stream {
-                admin: admin_addr,
-                recipient: Addr::unchecked("bob"),
-                balance: WrappedBalance::new_cw20(Addr::unchecked("cw20"), Uint128::new(750)),
-                claimed_balance: WrappedBalance::new_cw20(
-                    Addr::unchecked("cw20"),
-                    Uint128::new(250)
-                ),
-                start_time,
-                end_time,
-                title: None,
-                description: None,
-                paused: false,
-                paused_time: None,
-                paused_duration: None,
-                link_id: None,
-                is_detachable: true,
-            }
-        );
-
-        //Check final balances after distribution
-    }
-
-    #[test]
-    fn test_create_stream_with_refund() {
-        let mut deps = mock_dependencies();
-        let msg = InstantiateMsg { admin: None };
-        let info = mock_info("cw20", &[]);
-        instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let recipient = Addr::unchecked("bob");
-        let sender = info.sender.clone().into_string();
-        let sender_addr = info.sender.clone();
-
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: None,
-                recipient: recipient.into_string(),
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-
-        // Make sure remaining funds were refunded if duration didn't divide evenly into amount
-        let res = execute(deps.as_mut(), env, info, msg).unwrap();
-        let refund_msg = res.messages[0].clone().msg;
-        assert_eq!(
-            refund_msg,
-            CosmosMsg::Wasm(WasmMsg::Execute {
-                contract_addr: String::from("cw20"),
-                msg: to_binary(&Cw20ExecuteMsg::Transfer {
-                    recipient: sender,
-                    amount: Uint128::new(50)
-                })
-                .unwrap(),
-                funds: vec![]
-            })
-        );
-
-        let balance_amount = Uint128::new(350);
-        let balance = WrappedBalance::new_cw20(Addr::unchecked("cw20"), balance_amount);
-
-        assert_eq!(
-            get_stream(deps.as_ref(), 1).unwrap(),
-            Stream {
-                admin: sender_addr,
-                recipient: Addr::unchecked("bob"),
-                balance, // original amount - refund
-                claimed_balance: WrappedBalance::new_cw20(Addr::unchecked("cw20"), Uint128::new(0)),
-                start_time,
-                end_time,
-                title: None,
-                description: None,
-                paused: false,
-                paused_time: None,
-                paused_duration: None,
-                link_id: None,
-                is_detachable: true,
-            }
-        );
-    }
-    #[test]
-    fn test_execute_pause_stream() {
-        let mut deps = mock_dependencies();
-        let msg = InstantiateMsg { admin: None };
-        let sender = Addr::unchecked("alice").to_string();
-
-        let sender_addr = Addr::unchecked("alice");
-        let info = mock_info(&sender, &[]);
-        instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let stream_id: StreamId = 1;
-
-        let _ = execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::PauseStream { id: stream_id },
-        )
-        .unwrap();
-        let balance = WrappedBalance::new_cw20(Addr::unchecked("alice"), amount);
-        let saved_stream = get_stream(deps.as_ref(), stream_id).unwrap();
-        assert_eq!(
-            saved_stream,
-            Stream {
-                admin: Addr::unchecked("alice"),
-                recipient: Addr::unchecked("bob"),
-                balance, // original amount - refund
-                claimed_balance: WrappedBalance::new_cw20(sender_addr, Uint128::new(0)),
-                start_time,
-                end_time,
-                title: None,
-                description: None,
-                paused: true,
-                paused_time: saved_stream.paused_time,
-                paused_duration: saved_stream.paused_duration,
-                link_id: None,
-                is_detachable: true,
-            }
-        );
-    }
-    #[test]
-    fn test_invalid_start_time() {
-        let mut deps = mock_dependencies();
-
-        let msg = InstantiateMsg { admin: None };
-        let mut info = mock_info("alice", &[]);
-        instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let sender = Addr::unchecked("alice").to_string();
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(100);
-        let start_time = mock_env().block.time.plus_seconds(100).seconds();
-        let end_time = mock_env().block.time.plus_seconds(20).seconds();
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender,
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: None,
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        info.sender = Addr::unchecked("cw20");
-        let err = execute(deps.as_mut(), mock_env(), info, msg).unwrap_err();
-        assert_eq!(err, ContractError::InvalidStartTime {});
-    }
-
-    #[test]
-    fn invalid_cw20_addr() {
-        let mut deps = mock_dependencies();
-
-        let msg = InstantiateMsg { admin: None };
-        let mut info = mock_info("alice", &[]);
-        instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let sender = Addr::unchecked("alice").to_string();
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(100);
-        let start_time = mock_env().block.time.plus_seconds(100).seconds();
-        let end_time = mock_env().block.time.plus_seconds(200).seconds();
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender,
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: None,
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        info.sender = Addr::unchecked("wrongCw20");
-        let err = execute(deps.as_mut(), mock_env(), info, msg).unwrap_err();
-        assert_eq!(
-            err,
-            ContractError::Std(cosmwasm_std::StdError::GenericErr {
-                msg: "Invalid input: address not normalized".to_string()
-            })
-        );
-    }
-
-    #[test]
-    fn test_invalid_deposit_amount() {
-        let mut deps = mock_dependencies();
-
-        let msg = InstantiateMsg { admin: None };
-        let mut info = mock_info("alice", &[]);
-        instantiate(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let sender = Addr::unchecked("alice").to_string();
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(3);
-        let start_time = mock_env().block.time.plus_seconds(100).seconds();
-        let end_time = mock_env().block.time.plus_seconds(200).seconds();
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender,
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: None,
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        info.sender = Addr::unchecked("cw20");
-        let err = execute(deps.as_mut(), mock_env(), info, msg).unwrap_err();
-        assert_eq!(err, ContractError::AmountLessThanDuration {});
-    }
-    #[test]
-    fn test_execute_remove_stream() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-        let info = mock_info(&sender, &[]);
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let stream_id: StreamId = 1;
-
-        //Remove stream and verify not found error returned
-        let remove_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::RemoveStream { id: stream_id },
-        )
-        .unwrap();
-        let error = get_stream(deps.as_ref(), stream_id).unwrap_err();
-        assert_eq!(
-            error,
-            StdError::NotFound {
-                kind: "cw_payroll::state::Stream".to_string()
-            }
-        );
-
-        let refund_msg = remove_response.messages[0].clone().msg;
-        assert_eq!(
-            refund_msg,
-            CosmosMsg::Wasm(WasmMsg::Execute {
-                contract_addr: sender.clone(),
-                msg: to_binary(&Cw20ExecuteMsg::Transfer {
-                    recipient: sender,
-                    amount
-                })
-                .unwrap(),
-                funds: vec![]
-            })
-        );
-    }
-    #[test]
-    fn test_execute_link_stream_invalid() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-
-        let info = mock_info(&sender, &[]);
-
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let ids = vec![1, 2];
-
-        //Link stream and verify error returned
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::LinkStream { ids: ids.clone() },
-        )
-        .unwrap_err();
-        assert_eq!(
-            error,
-            ContractError::StreamNotFound {
-                stream_id: *ids.second().unwrap()
-            }
-        );
-
-        let ids = vec![1, 1];
-
-        //Link stream and verify error returned
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::LinkStream { ids },
-        )
-        .unwrap_err();
-        assert_eq!(error, ContractError::StreamsShouldNotBeEqual {});
-
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        let sender = Addr::unchecked("bob").to_string();
-        let ids = vec![1, 2];
-
-        let unauthorized_info = mock_info(&sender, &[]);
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            unauthorized_info,
-            ExecuteMsg::LinkStream { ids },
-        )
-        .unwrap_err();
-        assert_eq!(error, ContractError::Unauthorized {});
-    }
-
-    #[test]
-
-    fn test_execute_link_stream_valid() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-
-        let info = mock_info(&sender, &[]);
-
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        //Create stream 2
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-        let ids = vec![1, 2];
-        let response = execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::LinkStream { ids: ids.clone() },
-        )
-        .unwrap();
-
-        let left_stream = get_stream(deps.as_ref(), *ids.first().unwrap()).unwrap();
-        let right_stream = get_stream(deps.as_ref(), *ids.second().unwrap()).unwrap();
-        assert_eq!(left_stream.link_id, Some(*ids.second().unwrap()));
-        assert_eq!(right_stream.link_id, Some(*ids.first().unwrap()));
-        assert!(response
-            .attributes
-            .iter()
-            .any(|f| { f.key == "left_stream_id" }));
-        assert!(response
-            .attributes
-            .iter()
-            .any(|f| { f.key == "right_stream_id" }));
-    }
-    #[test]
-    fn test_execute_detach_stream_valid() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-
-        let info = mock_info(&sender, &[]);
-
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        //Create stream 2
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-        let ids = vec![1, 2];
-
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::LinkStream { ids: ids.clone() },
-        )
-        .unwrap();
-
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::DetachStream {
-                id: *ids.first().unwrap(),
-            },
-        )
-        .unwrap();
-        let left_stream = get_stream(deps.as_ref(), *ids.first().unwrap()).unwrap();
-        let right_stream = get_stream(deps.as_ref(), *ids.second().unwrap()).unwrap();
-
-        assert!(left_stream.paused);
-        assert!(left_stream.paused_time.is_some());
-        assert!(left_stream.link_id.is_none());
-
-        assert!(right_stream.paused);
-        assert!(right_stream.paused_time.is_some());
-        assert!(right_stream.link_id.is_none());
-    }
-
-    #[test]
-    fn test_execute_detach_stream_invalid() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-
-        let info = mock_info(&sender, &[]);
-
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: Some(false),
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        //Create stream 2
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: Some(false),
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-        let ids = vec![1, 2];
-
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::LinkStream { ids },
-        )
-        .unwrap();
-        let ids = vec![11, 22];
-
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::DetachStream {
-                id: *ids.first().unwrap(),
-            },
-        )
-        .unwrap_err();
-
-        assert_eq!(
-            error,
-            ContractError::StreamNotFound {
-                stream_id: *ids.first().unwrap()
-            }
-        );
-        let ids = vec![1, 22];
-
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::DetachStream {
-                id: *ids.second().unwrap(),
-            },
-        )
-        .unwrap_err();
-
-        assert_eq!(
-            error,
-            ContractError::StreamNotFound {
-                stream_id: *ids.second().unwrap()
-            }
-        );
-        let ids = vec![1, 2];
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::DetachStream {
-                id: *ids.second().unwrap(),
-            },
-        )
-        .unwrap_err();
-
-        assert_eq!(error, ContractError::StreamNotDetachable {});
-
-        let unauthorized_info = mock_info(&recipient, &[]);
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: Some(true),
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        //Create stream 2
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: Some(true),
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-        let ids = vec![3, 4];
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::LinkStream { ids: ids.clone() },
-        )
-        .unwrap();
-
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            unauthorized_info,
-            ExecuteMsg::DetachStream {
-                id: *ids.first().unwrap(),
-            },
-        )
-        .unwrap_err();
-
-        assert_eq!(error, ContractError::Unauthorized {});
-    }
-
-    #[test]
-    fn test_execute_resume_stream_valid() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-
-        let info = mock_info(&sender, &[]);
-
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        //Create stream 2
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient,
-                start_time,
-                end_time,
-                is_detachable: None,
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-        let ids = vec![1, 2];
-
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::LinkStream { ids: ids.clone() },
-        )
-        .unwrap();
-
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::DetachStream {
-                id: *ids.first().unwrap(),
-            },
-        )
-        .unwrap();
-        let left_stream = get_stream(deps.as_ref(), *ids.first().unwrap()).unwrap();
-        let right_stream = get_stream(deps.as_ref(), *ids.second().unwrap()).unwrap();
-
-        assert!(left_stream.paused);
-        assert!(left_stream.paused_time.is_some());
-        assert!(left_stream.link_id.is_none());
-
-        assert!(right_stream.paused);
-        assert!(right_stream.paused_time.is_some());
-        assert!(right_stream.link_id.is_none());
-
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::ResumeStream {
-                id: *ids.first().unwrap(),
-            },
-        )
-        .unwrap();
-        let left_stream = get_stream(deps.as_ref(), *ids.first().unwrap()).unwrap();
-        let right_stream = get_stream(deps.as_ref(), *ids.second().unwrap()).unwrap();
-
-        assert!(!left_stream.paused);
-        assert!(left_stream.paused_time.is_none());
-
-        assert!(right_stream.paused);
-        assert!(right_stream.paused_time.is_some());
-    }
-
-    #[test]
-    fn test_execute_resume_stream_invalid() {
-        let mut deps = mock_dependencies();
-        let sender = Addr::unchecked("alice").to_string();
-
-        let info = mock_info(&sender, &[]);
-
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg { admin: None },
-        )
-        .unwrap();
-
-        let recipient = Addr::unchecked("bob").to_string();
-        let amount = Uint128::new(350);
-        let env = mock_env();
-        let start_time = env.block.time.plus_seconds(100).seconds();
-        let end_time = env.block.time.plus_seconds(400).seconds();
-
-        //Create stream 1
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender.clone()),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: Some(true),
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        //Create stream 2
-        let msg = ExecuteMsg::Receive(Cw20ReceiveMsg {
-            sender: sender.clone(),
-            amount,
-            msg: to_binary(&ReceiveMsg::CreateStream {
-                admin: Some(sender),
-                recipient: recipient.clone(),
-                start_time,
-                end_time,
-                is_detachable: Some(false),
-            })
-            .unwrap(),
-        });
-        execute(deps.as_mut(), mock_env(), info.clone(), msg).unwrap();
-
-        let ids = vec![1, 2];
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            ExecuteMsg::LinkStream { ids },
-        )
-        .unwrap();
-        let ids = vec![1, 2];
-
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            ExecuteMsg::ResumeStream {
-                id: *ids.second().unwrap(),
-            },
-        )
-        .unwrap_err();
-        assert_eq!(error, ContractError::StreamNotPaused {});
-
-        let ids = vec![1, 2];
-        let unauthorized_info = mock_info(&recipient, &[]);
-        let error = execute(
-            deps.as_mut(),
-            mock_env(),
-            unauthorized_info,
-            ExecuteMsg::DetachStream {
-                id: *ids.first().unwrap(),
-            },
-        )
-        .unwrap_err();
-
-        assert_eq!(error, ContractError::StreamNotDetachable {});
-    }
 }
