+++ contracts/external/cw-payroll/src/lib.rs
@@ -1,7 +1,10 @@
-pub mod balance;
 pub mod contract;
 mod error;
 pub mod linking;
 pub mod msg;
 pub mod state;
+
 pub use crate::error::ContractError;
+
+#[cfg(test)]
+mod tests;
