+++ contracts/external/cw-payroll/src/msg.rs
@@ -1,39 +1,19 @@
 use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::Response;
+use cosmwasm_std::Uint128;
 use cw20::Cw20ReceiveMsg;
+use cw_denom::CheckedDenom;
 
-use crate::{balance::WrappedBalance, ContractError};
+use crate::state::{StreamId, StreamIds};
 
 #[cw_serde]
 pub struct InstantiateMsg {
     pub admin: Option<String>,
 }
-pub type StreamId = u64;
-pub type StreamIds = Vec<StreamId>;
-pub type ContractResult = Result<Response, ContractError>;
-
-pub(crate) trait StreamIdsExtensions {
-    fn second(&self) -> Option<&StreamId>;
-    fn validate(&self) -> Result<(), ContractError>;
-}
-impl StreamIdsExtensions for StreamIds {
-    fn second(&self) -> Option<&StreamId> {
-        self.get(1)
-    }
-    fn validate(&self) -> Result<(), ContractError> {
-        if self.len() != 2 {
-            return Err(ContractError::InvalidStreamIds {});
-        }
-        if self.first() == self.second() {
-            return Err(ContractError::StreamsShouldNotBeEqual {});
-        }
-        Ok(())
-    }
-}
 
 #[cw_serde]
 pub enum ExecuteMsg {
     Receive(Cw20ReceiveMsg),
+    Create {},
     Distribute {
         id: StreamId, // Stream id
     },
@@ -70,7 +50,8 @@ pub enum ReceiveMsg {
 pub struct StreamParams {
     pub admin: String,
     pub recipient: String,
-    pub balance: WrappedBalance,
+    pub balance: Uint128,
+    pub denom: CheckedDenom,
     pub start_time: u64,
     pub end_time: u64,
     pub title: Option<String>,
@@ -102,8 +83,9 @@ pub struct StreamResponse {
     pub id: u64,
     pub admin: String,
     pub recipient: String,
-    pub balance: WrappedBalance,
-    pub claimed_balance: WrappedBalance,
+    pub balance: Uint128,
+    pub claimed_balance: Uint128,
+    pub denom: CheckedDenom,
     pub start_time: u64,
     pub end_time: u64,
     pub paused_time: Option<u64>,
