+++ contracts/external/cw-payroll/src/state.rs
@@ -1,15 +1,15 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Addr, DepsMut, StdResult, Storage, Timestamp, Uint128};
+use cosmwasm_std::{Addr, Response, Timestamp, Uint128};
 use cw_storage_plus::{Item, Map};
 use serde::{Deserialize, Serialize};
 
-use crate::{balance::WrappedBalance, msg::StreamId};
+use crate::ContractError;
+use cw_denom::CheckedDenom;
 
 #[cw_serde]
 pub struct Config {
     pub admin: Addr,
 }
-
 pub const CONFIG: Item<Config> = Item::new("config");
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
@@ -17,8 +17,9 @@ pub struct Stream {
     pub admin: Addr,
     pub recipient: Addr,
     /// Balance in Native and Cw20 tokens
-    pub balance: WrappedBalance,
-    pub claimed_balance: WrappedBalance,
+    pub balance: Uint128,
+    pub claimed_balance: Uint128,
+    pub denom: CheckedDenom,
     pub start_time: u64,
     pub end_time: u64,
     pub paused_time: Option<u64>,
@@ -40,18 +41,19 @@ pub struct Stream {
 
 impl Stream {
     pub(crate) fn can_distribute_more(&self) -> bool {
-        if self.balance.amount() == 0 {
+        if self.balance == Uint128::zero() {
             return false;
         }
-        self.claimed_balance.amount() < self.balance.amount()
+        self.claimed_balance < self.balance
     }
+
     pub(crate) fn calc_distribution_rate_core(
         start_time: u64,
         end_time: u64,
         block_time: Timestamp,
         paused_duration: Option<u64>,
-        balance: &WrappedBalance,
-        claimed: &WrappedBalance,
+        balance: Uint128,
+        claimed_balance: Uint128,
     ) -> (Uint128, Uint128) {
         let block_time = std::cmp::min(block_time.seconds(), end_time);
         let duration: u64 = end_time.checked_sub(start_time).unwrap_or_default();
@@ -63,15 +65,12 @@ impl Stream {
                 .unwrap_or_default()
                 .into();
 
-            let rate_per_second = balance
-                .amount()
-                .checked_div(duration.into())
-                .unwrap_or_default();
+            let rate_per_second = balance.checked_div(duration.into()).unwrap_or_default();
 
             return (
                 Uint128::from(
-                    (passed * rate_per_second)
-                        .checked_sub(claimed.amount())
+                    (Uint128::from(passed) * rate_per_second)
+                        .checked_sub(claimed_balance)
                         .unwrap_or_default(),
                 ),
                 Uint128::from(rate_per_second),
@@ -79,16 +78,18 @@ impl Stream {
         }
         (Uint128::new(0), Uint128::new(0))
     }
+
     pub(crate) fn calc_distribution_rate(&self, block_time: Timestamp) -> (Uint128, Uint128) {
         Stream::calc_distribution_rate_core(
             self.start_time,
             self.end_time,
             block_time,
             self.paused_duration,
-            &self.balance,
-            &self.claimed_balance,
+            self.balance,
+            self.claimed_balance,
         )
     }
+
     pub(crate) fn calc_pause_duration(&self, block_time: Timestamp) -> Option<u64> {
         let end = std::cmp::min(block_time.seconds(), self.end_time);
         self.paused_duration.unwrap_or_default().checked_add(
@@ -97,26 +98,29 @@ impl Stream {
         )
     }
 }
+
+pub type StreamId = u64;
+pub type StreamIds = Vec<StreamId>;
+pub type ContractResult = Result<Response, ContractError>;
+
+pub(crate) trait StreamIdsExtensions {
+    fn second(&self) -> Option<&StreamId>;
+    fn validate(&self) -> Result<(), ContractError>;
+}
+impl StreamIdsExtensions for StreamIds {
+    fn second(&self) -> Option<&StreamId> {
+        self.get(1)
+    }
+    fn validate(&self) -> Result<(), ContractError> {
+        if self.len() != 2 {
+            return Err(ContractError::InvalidStreamIds {});
+        }
+        if self.first() == self.second() {
+            return Err(ContractError::StreamsShouldNotBeEqual {});
+        }
+        Ok(())
+    }
+}
+
 pub const STREAM_SEQ: Item<u64> = Item::new("stream_seq");
 pub const STREAMS: Map<StreamId, Stream> = Map::new("stream");
-
-pub fn add_stream(deps: DepsMut, stream: &Stream) -> StdResult<StreamId> {
-    let id = STREAM_SEQ.load(deps.storage)?;
-    let id = id.checked_add(1).unwrap();
-    STREAM_SEQ.save(deps.storage, &id)?;
-    STREAMS.save(deps.storage, id, stream)?;
-    Ok(id)
-}
-pub fn save_stream(
-    storage: &mut dyn Storage,
-    id: StreamId,
-    stream: &Stream,
-) -> StdResult<StreamId> {
-    STREAMS.save(storage, id, stream)?;
-    Ok(id)
-}
-
-pub fn remove_stream(storage: &mut dyn Storage, stream_id: StreamId) -> StdResult<StreamId> {
-    STREAMS.remove(storage, stream_id);
-    Ok(stream_id)
-}
